check for duplicate entries after skip logic, obviating allow_duplicates
As far as I understand it, allow_duplicates was added to support the test
case in fdroidserver where there are duplicate signature files in the test
APK.   It is a good idea to check for duplicate entries, since they are an
exploit vector, as long as the check does not include entries that would
be removed from the APK anyway.  This makes for tolerant inputs.